requirements.txt
ultralytics
torch
torchvision
Flask
Pillow
numpy
matplotlib
requests
